backend/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
PyMySQL>=1.1.0
DBUtils>=3.0.3
Faker>=19.0.0
faker-commerce>=1.0.3
flask-jwt-extended>=4.4.0
pandas>=2.0.0
azure-storage-queue>=12.14.0
azure-storage-blob>=12.17.0
azure-data-tables>=12.4.0
azure-core>=1.28.0
azure-storage-file-datalake>=12.14.0
